fix(#50): 캐시 삭제시 해당 로직에서 바로 저장할 필요가 없어서 분리한 cacheService를 삭제
