+++ service/product/product_server/src/main/java/com/sparta/product/application/CategoryService.java
@@ -5,9 +5,13 @@ import com.sparta.product.domain.repository.jpa.CategoryRepository;
 import com.sparta.product.presentation.exception.ProductErrorCode;
 import com.sparta.product.presentation.exception.ProductServerException;
 import com.sparta.product.presentation.response.CategoryResponse;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,9 +21,9 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j(topic = "CategoryService")
 public class CategoryService {
   private final CategoryRepository categoryRepository;
-  private final CategoryCacheService categoryCacheService;
 
   @Transactional
+  @CacheEvict(cacheNames = "categories-cache", key = "'categories'")
   public Long createCategory(String name, Long parentCategoryId) {
     Category parent =
         Optional.ofNullable(parentCategoryId).map(this::findByCategoryId).orElse(null);
@@ -27,11 +31,11 @@ public class CategoryService {
     Optional.ofNullable(parent).ifPresent(p -> p.addSubCategory(category));
 
     var saved = categoryRepository.save(category);
-    refreshCategoryCache();
     return saved.getCategoryId();
   }
 
   @Transactional
+  @CacheEvict(cacheNames = "categories-cache", key = "'categories'")
   public CategoryResponse updateCategory(
       Long targetCategoryId, String name, Long parentCategoryId) {
     Category target = findByCategoryId(targetCategoryId);
@@ -40,16 +44,22 @@ public class CategoryService {
 
     target.update(name, parent);
     syncParentCategory(target, parent); // 변경된 부모의 기존 연결 끊기
-    refreshCategoryCache();
-
     return CategoryResponse.fromEntity(target);
   }
 
   @Transactional
+  @CacheEvict(cacheNames = "categories-cache", key = "'categories'")
   public void deleteCategory(Long categoryId) {
     Category category = findByCategoryId(categoryId);
     categoryRepository.delete(category);
-    refreshCategoryCache();
+  }
+
+  @Cacheable(cacheNames = "categories-cache", key = "'categories'")
+  public List<CategoryResponse> fetchAndCacheCategories() {
+    return categoryRepository.findAllWithSubCategories().stream()
+        .filter(category -> category.getParent() == null)
+        .map(CategoryResponse::fromEntity)
+        .collect(Collectors.toList());
   }
 
   private void syncParentCategory(Category target, Category newParent) {
@@ -60,11 +70,6 @@ public class CategoryService {
     if (newParent != null) {
       newParent.addSubCategory(target);
     }
-  }
-
-  private void refreshCategoryCache() {
-    categoryCacheService.clearCache();
-    categoryCacheService.fetchAndCacheCategories();
   }
 
   private Category findByCategoryId(Long categoryId) {
